Add files via upload
The heartbeat 💓 of  CloneCraftedSite! This file features a stellar clone of a renowned website, blending HTML structure with CSS styling and JavaScript magic. Experience the smooth scroll 🚀 powered by Locomotive Scroll and the elegant touch of the Futura font. Dive into the code to witness the replication journey and personalized styling. Happy exploring! 🌐✨